+++ src/config/dictionary/parsing/features.mjs
@@ -287,6 +287,7 @@ export const PARSING_CHOICE_FEATURES = {
     "Training In War and Song",
     "Tools of the Trade",
     "Bestial Instincts",
+    "Purple Dragon Rook",
   ],
   NO_CHOICE_SECRET: [
     "Divine Order",
@@ -401,6 +402,7 @@ export const PARSING_CHOICE_FEATURES = {
     "Magic Item Plans",
     "Tools of the Trade",
     "Bestial Instincts",
+    "Purple Dragon Rook",
   ],
   OVERRIDE_CHOICE_FEATURE: [
     "Eldritch Invocations",