+++ src/parser/enrichers/class/artificer/SteelDefender.mjs
@@ -4,10 +4,26 @@ import DDBEnricherData from "../../data/DDBEnricherData.mjs";
 export default class SteelDefender extends DDBEnricherData {
 
   get activity() {
-    return {
-      noConsumeTargets: true,
-      noTemplate: true,
-    };
+    if (this.is2014) {
+      return {
+        noConsumeTargets: true,
+        noTemplate: true,
+      };
+    } else {
+      return {
+        targetType: "self",
+        noConsumeTargets: true,
+        noTemplate: true,
+        data: {
+          bonuses: {
+            "ac": "@abilities.int.mod",
+            "hp": "@classes.artificer.levels * 5",
+            "attackDamage": "@abilities.int.mod",
+            "healing": "@abilities.int.mod",
+          },
+        },
+      };
+    }
   }
 
   get override() {