+++ src/parser/enrichers/monster/SteelDefender/_module.mjs
@@ -1,0 +1,2 @@
+export { default as Repair } from "./Repair.mjs";
+export { default as SteelBond } from "./SteelBond.mjs";