+++ src/parser/enrichers/monster/_module.mjs
@@ -27,3 +27,4 @@ export * as Beholder from "./Beholder/_module.mjs";
 export * as BeastOfTheLand from "./BeastOfTheLand/_module.mjs";
 export * as BeastOfTheSea from "./BeastOfTheSea/_module.mjs";
 export * as BeastOfTheSky from "./BeastOfTheSky/_module.mjs";
+export * as SteelDefender from "./SteelDefender/_module.mjs";